Design adjustments: max slippage, wallet connect scan screen (#2798)
