+++ packages/core-mobile/app/new/features/swap/utils.ts
@@ -1,7 +1,7 @@
 export const isSlippageValid = (value: string): boolean => {
   return Boolean(
     (parseFloat(value) >= 0.1 &&
-      parseFloat(value) <= 100 &&
+      parseFloat(value) <= 50 &&
       value?.length <= 4) ||
       !value
   )
